get rid of the list view directive, it’s a pain to pass things into the directive since so much of the functionality relies on things already being on the scope
