--- src/ui/directives/queryResultsView/queryResultsView.html
@@ -1,2 +1,0 @@
-<div ng-include="viewSrc">
-</div>
