+++ src/ui/index.html
@@ -78,8 +78,6 @@
   <script src="./directives/codemirror/codemirror.js"></script>
   <script src="./directives/query/query.js"></script>
   <script src="./directives/query/queryCtrl.js"></script>
-  <script src="./directives/queryResultsView/queryResultsView.js"></script>
-  <script src="./directives/queryResultsView/queryResultsViewCtrl.js"></script>
   <script src="./directives/queryResultsExport/queryResultsExport.js"></script>
   <script src="./directives/queryResultsExport/queryResultsExportCtrl.js"></script>
   <script src="./directives/sidebar/sidebar.js"></script>
